% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folketingsvalg.R
\docType{data}
\name{folketingsvalg}
\alias{folketingsvalg}
\title{Folketingsvalg i Danmark}
\format{En data frame med 11 variabler: \code{Dato}, \code{Landsdel}, \code{Storkreds}, \code{Opstillingskreds}, \code{Afstemningsomraade},
  \code{Parti}, \code{Bogstav}, \code{Antal}, \code{Stemmeberettigede}, \code{DeltagelsePct}, \code{Person}}
\usage{
folketingsvalg
}
\description{
Et datasæt som indeholder resultaterne af folketingsvalgene 2011, 2015, 2019 og 2022.
}
\keyword{datasets}
